+++ src/declarations/ESTree.d.ts
@@ -4,6 +4,7 @@ import * as ESTree from 'estree';
 
 declare module 'estree' {
     interface BaseNode {
+        ignoredNode?: boolean;
         obfuscatedNode?: boolean;
         parentNode?: ESTree.Node;
     }
