+++ src/enums/container/node-transformers/NodeTransformer.ts
@@ -10,7 +10,9 @@ export enum NodeTransformer {
     LiteralTransformer,
     MemberExpressionTransformer,
     MethodDefinitionTransformer,
+    NodeGuardTransformer,
     ObjectExpressionTransformer,
+    ParentizeTransformer,
     TemplateLiteralTransformer,
     VariableDeclarationTransformer
 }
