+++ src/interfaces/node-transformers/INodeTransformer.d.ts
@@ -1,3 +1,4 @@
+import * as estraverse from 'estraverse';
 import * as ESTree from 'estree';
 
 import { IVisitor } from '../IVisitor';
@@ -9,9 +10,15 @@ export interface INodeTransformer {
     getVisitor (): IVisitor;
 
     /**
-     * @param node
-     * @param parentNode
-     * @returns {ESTree.Node}
+     * @param {Node} node
+     * @param {Node} parentNode
      */
-    transformNode (node: ESTree.Node, parentNode: ESTree.Node): ESTree.Node;
+    analyzeNode ? (node: ESTree.Node, parentNode: ESTree.Node): void;
+
+    /**
+     * @param {Node} node
+     * @param {Node} parentNode
+     * @returns {Node | VisitorOption}
+     */
+    transformNode (node: ESTree.Node, parentNode: ESTree.Node): ESTree.Node | estraverse.VisitorOption;
 }
