+++ src/node-guards/ConditionalCommentNodeGuard.ts
@@ -19,39 +19,39 @@ export class ConditionalCommentNodeGuard implements INodeGuard {
     /**
      * @type {boolean}
      */
-    private obfuscationEnabledForCurrentNode: boolean = true;
+    private obfuscationAllowedForCurrentNode: boolean = true;
 
     /**
      * @type {boolean}
      */
-    private obfuscationEnabledForNextNode: boolean | null = null;
+    private obfuscationAllowedForNextNode: boolean | null = null;
 
     /**
      * @returns {boolean}
      * @param node
      */
     public check (node: ESTree.Node): boolean {
-        if (this.obfuscationEnabledForNextNode) {
-            this.obfuscationEnabledForCurrentNode = this.obfuscationEnabledForNextNode;
-            this.obfuscationEnabledForNextNode = null;
+        if (this.obfuscationAllowedForNextNode) {
+            this.obfuscationAllowedForCurrentNode = this.obfuscationAllowedForNextNode;
+            this.obfuscationAllowedForNextNode = null;
         }
 
         if (!node.leadingComments && !node.trailingComments) {
-            return this.obfuscationEnabledForCurrentNode;
+            return this.obfuscationAllowedForCurrentNode;
         }
 
         const leadingComments: ESTree.Comment[] | undefined = node.leadingComments;
         const trailingComments: ESTree.Comment[] | undefined = node.trailingComments;
 
         if (leadingComments) {
-            this.obfuscationEnabledForCurrentNode = this.checkComments(leadingComments);
+            this.obfuscationAllowedForCurrentNode = this.checkComments(leadingComments);
         }
 
         if (trailingComments) {
-            this.obfuscationEnabledForNextNode = this.checkComments(trailingComments);
+            this.obfuscationAllowedForNextNode = this.checkComments(trailingComments);
         }
 
-        return this.obfuscationEnabledForCurrentNode;
+        return this.obfuscationAllowedForCurrentNode;
     }
 
     /**
@@ -61,22 +61,22 @@ export class ConditionalCommentNodeGuard implements INodeGuard {
     private checkComments (comments: ESTree.Comment[]): boolean {
         const commentsLength: number = comments.length;
 
-        let obfuscationEnabled: boolean = this.obfuscationEnabledForCurrentNode;
+        let obfuscationAllowed: boolean = this.obfuscationAllowedForCurrentNode;
 
         for (let i: number = 0; i < commentsLength; i++) {
             const comment: ESTree.Comment = comments[i];
 
             if (ConditionalCommentNodeGuard.obfuscationEnableCommentRegExp.test(comment.value)) {
-                obfuscationEnabled = true;
+                obfuscationAllowed = true;
 
                 continue;
             }
 
             if (ConditionalCommentNodeGuard.obfuscationDisableCommentRegExp.test(comment.value)) {
-                obfuscationEnabled = false;
+                obfuscationAllowed = false;
             }
         }
 
-        return obfuscationEnabled;
+        return obfuscationAllowed;
     }
 }
