+++ test/dev/dev.ts
@@ -11,6 +11,7 @@ import { NO_CUSTOM_NODES_PRESET } from '../../src/options/presets/NoCustomNodes'
                 var foo = function () {
                     console.log('abc');
                 };
+                // javascript-obfuscator:disable
                 var bar = function () {
                     console.log('def');
                 };
