+++ test/functional-tests/custom-nodes/domain-lock-nodes/DomainLockNode.spec.ts
@@ -6,7 +6,7 @@ import { NO_CUSTOM_NODES_PRESET } from '../../../../src/options/presets/NoCustom
 
 import { readFileAsString } from '../../../helpers/readFileAsString';
 
-import { JavaScriptObfuscator } from '../../../../src/JavaScriptObfuscator';
+import { JavaScriptObfuscator } from '../../../../src/JavaScriptObfuscatorFacade';
 
 describe('DomainLockNode', () => {
     const regExp: RegExp = /var _0x([a-f0-9]){4,6} *= *new RegExp/;
