+++ test/functional-tests/custom-nodes/string-array-nodes/StringArrayRotateFunctionNode.spec.ts
@@ -6,7 +6,7 @@ import { NO_CUSTOM_NODES_PRESET } from '../../../../src/options/presets/NoCustom
 
 import { readFileAsString } from '../../../helpers/readFileAsString';
 
-import { JavaScriptObfuscator } from '../../../../src/JavaScriptObfuscator';
+import { JavaScriptObfuscator } from '../../../../src/JavaScriptObfuscatorFacade';
 
 describe('StringArrayRotateFunctionNode', () => {
     const regExp: RegExp = /while *\(-- *_0x([a-f0-9]){4,6}\) *\{/;
